+++ ChangeLog.md
@@ -2,7 +2,10 @@
 
 ## v0.12.0 (in development)
 
-No changes yet.
+### Enhancements
+
+ * New command, `erlang set-in-tool-versions`, updates the [`tool-versions` file](https://asdf-vm.com/manage/configuration.html#tool-versions) with the specified Erlang version
+ * `releases install latest` now fetches the latest GA release from GitHub when no versions are installed locally
 
 
 ## v0.11.0 (Jan 26, 2026)